feat(ma): MA directives implementation + security hardening + full code review fixes
Phase 1: MA 6대 지시사항 구현
- PEF 레지스트리 모델 + FI 자동 추천 API (deal_value 기반 GP 매칭)
- Short-List 승격 체계 (is_short_listed 플래그 + 연락처 필수 검증)
- 입찰 결과 추적 (BID_SUBMITTED/NOT_SUBMITTED/DROPPED)
- 파이프라인 9단계 재구성 (BIDDING_DD→BIDDING + MOU_SIGNED + MAIN_DUE_DILIGENCE)
- 계약서 생성 모듈 (템플릿 + 변수 바인딩 + 조항 관리)

Phase 2: 통합 코드 리뷰 17개 Critical 이슈 수정 (13건 적용)
- SEC-001: VDR 내부키 빈문자열 바이패스 차단
- SEC-002: Ralph API 경로 순회(Path Traversal) 방지
- SEC-003: AUTH_ENABLED 화이트리스트 전환 (블랙리스트→화이트리스트)
- C-01: 입찰금액 float→Decimal 정밀도 보장
- C-02: BlobStorage sync I/O→asyncio.to_thread 전환
- P1: RFI sync N+1 쿼리 제거 (IN 쿼리 + 인메모리 처리)
- P2: document_extraction 활성 추출 IN 쿼리 최적화
- P3: SI mapping 페이지네이션 한도 (10,000건)
- P4: SI mapping 4개 COUNT→단일 집계 쿼리
- P5: buyers API 페이지네이션 추가 (limit/offset)
- M-C01: AuditAction enum 마이그레이션 051 추가
- M-C04: 046 마이그레이션 autocommit_block 전환
- M-C05: 046 downgrade 데이터 유실 방지 가드

기타: IM/KIIS 린트 정리, 리뷰 시스템 rules→skills 재배치, KIIS GP 펀드 링크 테스트

검증: ruff 0 errors, pytest 1,146 passed / 14 skipped / 0 failed

+++ deal-mgmt/app/schemas/dashboard.py
@@ -1,18 +1,20 @@
 from __future__ import annotations
 
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class PhaseSummary(BaseModel):
     phase: str
     count: int
-    total_value: float | None = None
+    total_value: Decimal | None = None
 
 
 class DashboardStats(BaseModel):
     total_transactions: int
     active_transactions: int
-    total_deal_value: float | None = None
+    total_deal_value: Decimal | None = None
     by_phase: list[PhaseSummary]
     by_status: dict[str, int]
     by_side: dict[str, int]
